refactor kubernetes trainer
