+++ tfx/extensions/experimental/kubernetes/__init__.py
@@ -1,4 +1,5 @@
-# Copyright 2020 Google LLC. All Rights Reserved.
+# Lint as: python2, python3
+# Copyright 2019 Google LLC. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
